move license page to correct location
